Move test_basic utilities to utils.py and SciPy tests to test_basic_scipy
