+++ .gitignore
@@ -40,3 +40,4 @@ Theano.suo
 .ropeproject
 core
 .idea
+/htmlcov/
