+++ doc/extending/unittest.txt
@@ -425,7 +425,7 @@ verify that the op generates the proper output, that the gradient is
 valid, that the Op fails in known/expected ways. Because so much of
 this is common, two helper functions exists to make your lives easier:
 ``makeTester`` and ``makeBroadcastTester`` (defined in module
-``tests.tensor.test_basic``).
+``tests.tensor.utils``).
 
 Here is an example of ``makeTester`` generating testcases for the Dot
 product op:
@@ -435,7 +435,7 @@ product op:
     from numpy import dot
     from numpy.random import rand
 
-    from tests.tensor.test_basic import makeTester
+    from tests.tensor.utils import makeTester
 
     TestDot = makeTester(name = 'DotTester',
                          op = dot,
