+++ setup.cfg
@@ -5,6 +5,7 @@ max-line-length = 88
 per-file-ignores =
     **/__init__.py:F401,E402,F403
     theano/sparse/sandbox/sp2.py:F401
+    tests/tensor/test_basic_scipy.py:E402
     tests/sparse/test_basic.py:E402
     tests/sparse/test_opt.py:E402
     tests/sparse/test_sp2.py:E402
