+++ tests/gpuarray/test_basic_ops.py
@@ -1,12 +1,10 @@
-import pytest
-
-
-pygpu = pytest.importorskip("pygpu")
-gpuarray = pygpu.gpuarray
 import numpy as np
+import pytest
 
 import theano
 import theano.tensor as tt
+
+# Don't import test classes otherwise they get tested as part of the file
 from tests import unittest_tools as utt
 from tests.gpuarray.config import mode_with_gpu, mode_without_gpu, test_ctx_name
 from tests.tensor.test_basic import (
@@ -14,9 +12,8 @@ from tests.tensor.test_basic import (
     TestComparison,
     TestJoinAndSplit,
     TestReshape,
-    rand,
-    safe_make_node,
 )
+from tests.tensor.utils import rand, safe_make_node
 from theano.gpuarray.basic_ops import (
     GpuAlloc,
     GpuAllocEmpty,
@@ -35,12 +32,13 @@ from theano.gpuarray.basic_ops import (
 )
 from theano.gpuarray.elemwise import GpuDimShuffle, GpuElemwise
 from theano.gpuarray.subtensor import GpuSubtensor
-
-# Don't import test classes otherwise they get tested as part of the file
 from theano.gpuarray.type import GpuArrayType, get_context, gpuarray_shared_constructor
 from theano.tensor import TensorType
 from theano.tensor.basic import alloc
 
+
+pygpu = pytest.importorskip("pygpu")
+gpuarray = pygpu.gpuarray
 
 utt.seed_rng()
 rng = np.random.RandomState(seed=utt.fetch_seed())
