+++ tests/tensor/nnet/test_nnet.py
@@ -4,7 +4,7 @@ import pytest
 import theano
 import theano.tensor as tt
 from tests import unittest_tools as utt
-from tests.tensor.test_basic import (
+from tests.tensor.utils import (
     _good_broadcast_unary_normal_float_no_complex,
     check_floatX,
     makeBroadcastTester,
