+++ tests/tensor/test_blas.py
@@ -22,8 +22,9 @@ import theano
 import theano.tensor as tt
 import theano.tensor.blas_scipy
 from tests import unittest_tools
-from tests.tensor.test_basic import as_tensor_variable, compile, inplace, inplace_func
+from tests.tensor.utils import inplace_func
 from theano import In, config, shared
+from theano.tensor import as_tensor_variable, inplace
 from theano.tensor.blas import (
     Dot22,
     Dot22Scalar,
@@ -109,7 +110,7 @@ class TestGemm:
                 f = inplace_func(
                     [tz, ta, tx, ty, tb],
                     gemm_inplace(tz, ta, tx, ty, tb),
-                    mode=compile.Mode(optimizer=None, linker=l),
+                    mode=theano.compile.Mode(optimizer=None, linker=l),
                 )
                 f(z, a, x, y, b)
                 z_after = self._gemm(z_orig, a, x, y, b)
@@ -274,12 +275,12 @@ class TestGemm:
             tz, ta, tx, ty, tb = [shared(p) for p in (z, a, x, y, b)]
 
             # f = inplace_func([tz,ta,tx,ty,tb], gemm_inplace(tz,ta,tx,ty,tb),
-            #                 mode = compile.Mode(optimizer = None, linker=l))
+            #                 mode = theano.compile.Mode(optimizer = None, linker=l))
             # f(z, a, x, y, b)
             f = inplace_func(
                 [],
                 gemm_inplace(tz, ta, tx, ty, tb),
-                mode=compile.Mode(optimizer=None, linker=l),
+                mode=theano.compile.Mode(optimizer=None, linker=l),
             )
             f()
             unittest_tools.assert_allclose(z_after, tz.get_value(borrow=True))
@@ -336,7 +337,7 @@ class TestGemm:
                 f_i = inplace_func(
                     [],
                     gemm_inplace(tz[:, :, i], ta, tx[:, :, i], ty[:, :, i], tb),
-                    mode=compile.Mode(optimizer=None, linker=l),
+                    mode=theano.compile.Mode(optimizer=None, linker=l),
                 )
                 for j in range(3):
                     # tz will not _always_ be overwritten,
@@ -355,7 +356,7 @@ class TestGemm:
                     [],
                     tz_i,
                     updates=[(tz, tt.set_subtensor(tz[:, :, i], tz_i))],
-                    mode=compile.Mode(optimizer=None, linker=l),
+                    mode=theano.compile.Mode(optimizer=None, linker=l),
                 )
                 for j in range(3):
                     g_i()
@@ -612,7 +613,7 @@ def just_gemm(i, o, ishapes=None, max_graphlen=0, expected_nb_gemm=1):
     g = inplace_func(
         i,
         o,
-        mode=compile.Mode(linker="py", optimizer=None),
+        mode=theano.compile.Mode(linker="py", optimizer=None),
         allow_input_downcast=True,
         on_unused_input="ignore",
     )
@@ -701,7 +702,7 @@ def test_gemm_opt_double_gemm():
     g = inplace_func(
         i,
         o,
-        mode=compile.Mode(linker="py", optimizer=None),
+        mode=theano.compile.Mode(linker="py", optimizer=None),
         on_unused_input="ignore",
     )
